feat: add getTier
